+++ pkg/mads/reconcile/interfaces.go
@@ -1,6 +1,8 @@
 package reconcile
 
 import (
+	"context"
+
 	envoy_core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
 
 	util_xds "github.com/Kong/kuma/pkg/util/xds"
@@ -8,10 +10,10 @@ import (
 
 // Reconciler re-computes configuration for a given node.
 type Reconciler interface {
-	Reconcile(*envoy_core.Node) error
+	Reconcile(context.Context, *envoy_core.Node) error
 }
 
 // Generates a snapshot of xDS resources for a given node.
 type SnapshotGenerator interface {
-	GenerateSnapshot(*envoy_core.Node) (util_xds.Snapshot, error)
+	GenerateSnapshot(context.Context, *envoy_core.Node) (util_xds.Snapshot, error)
 }