+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Changes:
 
+* feature: generate MonitoringAssignment for each Dataplane in a Mesh
+  [#532](https://github.com/Kong/kuma/pull/532)
 * feature: add a Monitoring Assignment Discovery Service (MADS) server
   [#531](https://github.com/Kong/kuma/pull/531)
 * feature: add a generic watchdog for xDS streams